update to `Fly.source`
Currently the docs say you can do `source(...globs)` but it doesn't work unless you wrap them in an array. This fix allows you do do what the globs say.

Also added the ability to pass in options to the source that are passed to `globby` which are passed to `node-glob`.