Update to `Fly.prototype.source`
This updates `Fly.prototype.source` to work like the documentation shows where you can pass in multiple files in `arguments` format. Previously you could only pass in 1 argument that was either a string or array which was miss leading in the documentation of this function.

The other thing that was updated was to go ahead and expand all the globs in the source instead of everytime unwrap is called. This should solve [178](https://github.com/flyjs/fly/issues/178).