Remember globs; Add start/watch options (#189)
* Update to `Fly.prototype.source`

This updates `Fly.prototype.source` to work like the documentation shows where you can pass in multiple files in `arguments` format. Previously you could only pass in 1 argument that was either a string or array which was miss leading in the documentation of this function.

The other thing that was updated was to go ahead and expand all the globs in the source instead of everytime unwrap is called. This should solve [178](https://github.com/flyjs/fly/issues/178).

* allow `exec` to pass in an array of arguments

* Update to `Fly.prototype.clear`

This update makes the clear function work just like it's documented.

* Update `Fly.prototype.target`

This update just updates the way globs are looped over since the `Fly.prototype.source` was updated to go ahead parse the globs prior to target function running.

* Update `Fly.prototype.watch`

Add the functionality to pass in the file that has changed into task that's executed. This way tasks have the option of passing in only the file that was changed.

This update solves [187](https://github.com/flyjs/fly/issues/187)

* Update `Fly.prototype.start`

This update is to be able to pass in arguments in the same way as `source ` and `clear` do.

* Documented the new watch functionality